move tests from unit to integration tests run tests in one test method because when multiple threads access the store and update data the tests are not valid anymore
