+++ tests/common/mod.rs
@@ -1,5 +1,6 @@
 use std::path::Path;
 use sysrepo::connection::{ConnectionOptions, SrConnection};
+use sysrepo::enums::SrDatastore;
 
 const YANG: &str = "./assets/yang";
 const TEST_MODULE: &str = "./assets/yang/test_module.yang";
@@ -9,24 +10,35 @@ const SUB: &str = "./assets/yang/sub.yang";
 
 pub struct Setup {
     connection: SrConnection,
-    modules: Vec<String>,
 }
 
 impl Setup {
     pub fn setup_test_module() -> Self {
-        let connection = SrConnection::new(ConnectionOptions::Datastore_Running).unwrap();
-        match connection.remove_module("test_module", true) {
-            Ok(_) => (),
-            Err(e) => println!("Setup Test Module: {e}"),
-        }
-
+        let mut connection = SrConnection::new(ConnectionOptions::Datastore_Running).unwrap();
         connection
             .install_module(Path::new(TEST_MODULE), None, None)
             .unwrap();
         let modules = vec!["test_module".to_string()];
-        Self {
-            connection,
-            modules,
+        let stores = vec![
+            SrDatastore::Startup,
+            SrDatastore::Running,
+            SrDatastore::Candidate,
+        ];
+        for store in stores {
+            let session = connection.start_session(store).unwrap();
+            session.replace_config(None, None, None).unwrap();
         }
+        Self { connection }
+    }
+
+    pub fn setup_example() -> Self {
+        let con = SrConnection::new(ConnectionOptions::Datastore_Running).unwrap();
+        con.install_module(
+            Path::new("./assets/yang/examples@2017-01-19.yang"),
+            None,
+            None,
+        )
+        .unwrap();
+        Self { connection: con }
     }
 }
